# backend/openapi.yaml
openapi: 3.0.3
info:
  title: HR Analytics API
  version: "0.1.0"
servers:
  - url: http://localhost:4000
paths:
  /api/uploads/excel:
    post:
      summary: Upload HR Excel (template or vendor report)
      operationId: uploadExcel
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          $ref: "#/components/responses/UploadAccepted"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/uploads/{uploadId}/status:
    get:
      summary: Check upload processing status
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/UploadStatus"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    UploadSummary:
      type: object
      properties:
        uploadId:
          type: string
        success:
          type: boolean
        sheets:
          type: array
          items:
            type: object
            properties:
              sheet:
                type: string
              processed_rows:
                type: integer
              failed_rows:
                type: integer
...
  responses:
    UploadAccepted:
      description: Upload queued for processing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UploadSummary"
    UploadStatus:
      description: Upload processing status
      content:
        application/json:
          schema:
            type: object
            properties:
              uploadId:
                type: string
              status:
                type: string
              processed_rows:
                type: integer
              errors:
                type: array
                items:
                  type: object
                  properties:
                    sheet:
                      type: string
                    row:
                      type: integer
                    column:
                      type: string
                    message:
                      type: string
...
    ValidationError:
      description: Validation failure
      content:
        application/json:
          schema:
            type: object
            required: [success, errors]
            properties:
              success:
                type: boolean
              errors:
                type: array
                items:
                  type: object
                  properties:
                    sheet:
                      type: string
                    row:
                      type: integer
                    column:
                      type: string
                    message:
                      type: string